// packages/design_system/components/button/src/interfaces/IButtonStateStyleProps.tsx
import styled, {css, keyframes} from 'styled-components'
    
interface ButtonStateStyleProps {
    buttonStateWidth?: string;
    borderRadius?: string;
    buttonStyle?: string;
    elevation?: string;
    mt?: string;
    mr?: string;
    mb?: string;
    ml?: string;
    type?: string;
    background?: string;
}
export type{
    ButtonStateStyleProps
}